バグううう

+++ WinOGL/CVertex.h
@@ -18,7 +18,7 @@ public:
 	//　頂点のY座標を書き込む
 	void SetY(double new_y);
 	//　頂点のX・Y座標の両方を書き込む
-	void SetXY(int new_x, int new_y);
+	void SetXY(double new_x, double new_y);
 	//　頂点のX座標を読み込む込む
 	double GetX();
 	//　頂点のY座標を読み込む
